backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
python-jose[cryptography]
supabase